backend/requirements.txt
Flask
flask-cors
google-cloud-storage
google-cloud-vision
azure-cognitiveservices-vision-computervision
Pillow
torch
torchvision
tqdm
requests
ebaysdk
openai
pandas
numpy
scikit-image
python-dotenv
rich
transformers
deepspeed
